StackRemaining message are changed to floats

+++ firmware/shared/src/io/Io_SharedStackWaterMark.c
@@ -42,11 +42,11 @@ void Io_SharedStackWaterMark_Check(struct stack_watermark *stacks, size_t num_of
     }
 }
 
-static uint8_t Io_StackSize(TaskHandle_t xTask, uint32_t stack_size, float watermark_threshold)
+static float Io_StackSize(TaskHandle_t xTask, uint32_t stack_size, float watermark_threshold)
 {
     assert((0.0f < watermark_threshold) && (watermark_threshold < 1.0f));
 
-    uint8_t stack_high_watermark_percent_remaining = (uint8_t)uxTaskGetStackHighWaterMark(xTask) / (uint8_t)stack_size;
+    float stack_high_watermark_percent_remaining = (float)uxTaskGetStackHighWaterMark(xTask) / (float)stack_size;
 
     return stack_high_watermark_percent_remaining;
 }
